# feat: optimize pre-commit configuration with modern best practices