## Summary
- Add check-added-large-files hook with 10MB limit to prevent large file commits
- Switch from 'script' to 'system' language for better compatibility  
- Add verbose output for improved debugging experience
- Optimize pre-push to use quick-only mode for faster feedback (~15s vs full CI)
- Enhance configuration documentation and structure
- Fix trailing whitespace and end-of-file issues in workflow files

## Test plan
- [x] Pre-commit hooks tested and working (format + lint checks in ~1.5s)
- [x] Pre-push hooks tested and working (format + lint + types in ~15s) 
- [x] Large file prevention tested with 10MB limit
- [x] All GitHub workflow files cleaned up (trailing whitespace, EOF)
- [x] CI workflow passed successfully

## Performance improvements
- Pre-commit: Fast format/lint feedback (~1.5 seconds)
- Pre-push: Balanced checks without full test suite (~15 seconds vs several minutes)
- Better error reporting with verbose output
- Prevention of accidental large file commits

The configuration now provides optimal developer experience with modern best practices.